+++ gateway/routes/services_routs.py
@@ -1,0 +1,10 @@
+USER_URL = 'http://user:5002'
+CONTROL_URL = 'http://controller:5001'
+STRATEGIES_URL = 'http://strategies:5003'
+
+# Para testes locais, descomente as linhas abaixo e comente as de cima
+# USER_URL = 'http://localhost:5002'
+# CONTROL_URL = 'http://localhost:5001'
+# STRATEGIES_URL = 'http://localhost:5003'
+
+# SECRET_KEY = "sua_chave_super_secreta"
