+++ gateway/templates/control/list_all_sessions.html
@@ -48,6 +48,7 @@
                         {% else %}
                         <em>No students</em>
                         {% endif %}
+                        <a href="/sessions/{{ session.id }}" class="btn btn-outline-primary mt-3 w-100">Ver Detalhes</a>
                     </div>
                 </div>
             </div>
