+++ gateway/templates/control/show_session.html
@@ -1,0 +1,71 @@
+<!DOCTYPE html>
+<html lang="en">
+
+<head>
+    <meta charset="UTF-8">
+    <title>Session Details</title>
+    <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.0/dist/css/bootstrap.min.css" rel="stylesheet">
+</head>
+
+<body class="bg-light">
+    <div class="container mt-5">
+        <div class="card shadow-lg">
+            <div class="card-header bg-primary text-white">
+                <h4 class="mb-0">Session #{{ session.id }}</h4>
+            </div>
+            <div class="card-body">
+                <h5 class="card-title">Status:
+                    <span class="badge 
+                        {% if session.status == 'active' %}
+                            bg-success
+                        {% elif session.status == 'pending' %}
+                            bg-warning text-dark
+                        {% else %}
+                            bg-secondary
+                        {% endif %}">
+                        {{ session.status.capitalize() }}
+                    </span>
+                </h5>
+
+                <hr>
+
+                <div class="mb-3">
+                    <h6>Strategies</h6>
+                    {% if session.strategies %}
+                    {% for strategy in session.strategies %}
+                    <span class="badge bg-info text-dark me-1">{{ strategy }}</span>
+                    {% endfor %}
+                    {% else %}
+                    <em>No strategies assigned.</em>
+                    {% endif %}
+                </div>
+
+                <div class="mb-3">
+                    <h6>Teachers</h6>
+                    {% if session.teachers %}
+                    {% for teacher in session.teachers %}
+                    <span class="badge bg-success me-1">{{ teacher }}</span>
+                    {% endfor %}
+                    {% else %}
+                    <em>No teachers assigned.</em>
+                    {% endif %}
+                </div>
+
+                <div class="mb-3">
+                    <h6>Students</h6>
+                    {% if session.students %}
+                    {% for student in session.students %}
+                    <span class="badge bg-secondary me-1">{{ student }}</span>
+                    {% endfor %}
+                    {% else %}
+                    <em>No students assigned.</em>
+                    {% endif %}
+                </div>
+
+                <a href="/sessions" class="btn btn-outline-primary mt-3">← Back to Sessions</a>
+            </div>
+        </div>
+    </div>
+</body>
+
+</html>
